+++ src/TPCTrainco.Umbraco.Extensions/ViewModels/Backbone/Sch.cs
@@ -8,8 +8,8 @@ namespace TPCTrainco.Umbraco.Extensions.ViewModels.Backbone
 {
     public class Sch
     {
-        public int Id { get; set; }
-        public int LocationId { get; set; }
+        public long Id { get; set; }
+        public long LocationId { get; set; }
         public int CourseId { get; set; }
         public int CityId { get; set; }
         public string DaysTitle { get; set; }
