+++ src/TPCTrainco.Umbraco.Extensions/ViewModels/Seminar.cs
@@ -8,7 +8,7 @@ namespace TPCTrainco.Umbraco.Extensions.ViewModels
 {
     public class Seminar
     {
-        public int SeminarId { get; set; }
+        public long SeminarId { get; set; }
         public int CourseId { get; set; }
         public string Title { get; set; }
         public string SubTitle { get; set; }
